move user specified info to new file
